requirements.txt
aiofiles>=22.1.0
aiohttp==3.8.4
apscheduler==3.6.3
alphabet-detector==0.0.7
asyncio==3.4.3
beautifulsoup4==4.12.2
bing_image_downloader==1.1.2
bleach==6.0.0
bs4==0.0.1
cloudscraper==1.2.71
cryptography==40.0.2
cryptg==0.4.0
cachetools==4.2.2
envparse-stubs==0.2.2
emoji==2.2.0
ffmpeg-python==0.2.0
future==0.18.3
glitch_this==1.0.2
geopy==2.3.0
gpytranslate==1.5.1
gtts==2.3.2
hachoir==3.2.0
httpx==0.24.0
httpx[http2]
httpcore==0.17.0
humanize==4.6.0
jikanpy==4.3.2
lxml==4.9.2
markdown==3.4.3
markdown2==2.4.8
motor==3.1.2
multicolorcaptcha==1.2.0
nekos.py==1.1.0
numpy==1.24.3
odmantic==0.9.2
opencv-python-headless==4.8.0.74
pillow==9.5.0
psutil==5.9.5
psycopg2-binary==2.9.6
PyDictionary==2.0.1
pykeyboard==0.1.5
python-arq==6.0.5
pymongo[srv]==4.3.3
pyjokes==0.6.0
pyyaml==6.0
pyrate-limiter==2.10.0
pynewtonmath==1.0.1
pyrogram==2.0.106
python-dotenv==1.0.0
python-telegram-bot==13.15
pytimeparse==1.1.8
pydantic==2.0.2
redis==4.5.5
regex==2023.5.5
requests==2.30.0
search_engine_parser==0.6.8
secureme==0.0.4
spamwatch==0.3.0
speedtest-cli==2.1.3
sqlalchemy==1.4.47
telegraph==2.2.0
telethon==1.28.5
tgcrypto==1.2.5
ujson==5.7.0
uvloop==0.17.0
wget==3.2
